requirements.txt
asgiref
dj-rest-auth
Django
django-cors-headers
django-environ
django-filter
django-rest-authtoken
djangorestframework
djangorestframework-simplejwt
environ
ldap3
Markdown
pillow
pyasn1
pydirectory
PyJWT
python-dotenv
sqlparse
tzdata=
